+++ LogixEngine/Code/Registry/IIdentifiable.cs
@@ -1,0 +1,11 @@
+﻿namespace LogixEngine.Registry
+{
+	/* ---------------------------------------------------------------------------- *
+     * CLASS.cs created by DrLogiq on DATE.
+     * Copyright © DrLogiq. All rights reserved.
+     * ---------------------------------------------------------------------------- */
+	public interface IIdentifiable
+	{
+		identifier Identifier { get; }
+	}
+}