+++ Terrascape/Changes.md
@@ -2,6 +2,14 @@
 For tracking changes between commits. Intended only for the developer.
 In other words, if you are not Logiq, ignore this.
 
-### Initial tasks completed:
+## Tasks to complete:
+- (None)
 
-- Tested the Debug.Profile(...) system
+## Completed:
+
+New:
+- [x] Tested the ability to register custom types
+- [x] Printed the profiler granularity, testing a new method of number -> string formatting
+
+Old:
+- [x] Tested the Debug.Profile(...) system